refactor: 패키지 구조 변경
- 서비스 domain 과 서비스 전반적으로 사용되는 global 기능들을 구분
- global 기능은 global 패키지에 작성한다.
