--- src/main/java/com/whatpl/exception/BizException.java
@@ -1,19 +1,0 @@
-package com.whatpl.exception;
-
-import lombok.Getter;
-
-@Getter
-public class BizException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-
-    public BizException(ErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-
-    public BizException(ErrorCode errorCode, Throwable cause) {
-        super(errorCode.getMessage(), cause);
-        this.errorCode = errorCode;
-    }
-}
