--- src/main/java/com/whatpl/exception/ExceptionControllerAdvice.java
@@ -1,25 +1,0 @@
-package com.whatpl.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class ExceptionControllerAdvice {
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleException(Exception e) {
-        log.error("handleException", e);
-        ErrorResponse errorResponse = ErrorResponse.of(ErrorCode.GLOBAL_EXCEPTION);
-        return new ResponseEntity<>(errorResponse, errorResponse.getStatus());
-    }
-
-    @ExceptionHandler(BizException.class)
-    public ResponseEntity<ErrorResponse> handleCustomException(BizException e) {
-        log.error("handleCustomException", e);
-        ErrorResponse errorResponse = ErrorResponse.of(e.getErrorCode());
-        return new ResponseEntity<>(errorResponse, errorResponse.getStatus());
-    }
-}
