+++ src/main/java/com/whatpl/global/security/handler/LoginFailureHandler.java
@@ -1,0 +1,31 @@
+package com.whatpl.global.security.handler;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.whatpl.global.exception.ErrorCode;
+import com.whatpl.global.exception.ErrorResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+@RequiredArgsConstructor
+public class LoginFailureHandler implements AuthenticationFailureHandler {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
+        response.setStatus(HttpStatus.BAD_REQUEST.value());
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
+        ErrorResponse errorResponse = ErrorResponse.of(ErrorCode.LOGIN_FAILED);
+        objectMapper.writeValue(response.getWriter(), errorResponse);
+    }
+}
