+++ src/main/java/com/whatpl/global/security/handler/LoginSuccessHandler.java
@@ -1,0 +1,40 @@
+package com.whatpl.global.security.handler;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.whatpl.global.security.domain.AccountPrincipal;
+import com.whatpl.global.jwt.JwtResponse;
+import com.whatpl.global.jwt.JwtService;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+@RequiredArgsConstructor
+public class LoginSuccessHandler implements AuthenticationSuccessHandler {
+
+    private final ObjectMapper objectMapper;
+    private final JwtService jwtService;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
+        AccountPrincipal principal = (AccountPrincipal) authentication.getPrincipal();
+        String accessToken = jwtService.createAccessToken(authentication);
+        String refreshToken = jwtService.createRefreshToken(principal.getId());
+        JwtResponse tokenResponse = JwtResponse.builder()
+                .accessToken(accessToken)
+                .refreshToken(refreshToken)
+                .build();
+
+        response.setStatus(HttpStatus.OK.value());
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
+        objectMapper.writeValue(response.getWriter(), tokenResponse);
+    }
+}
