+++ src/main/java/com/whatpl/global/security/handler/NoAuthorizationHandler.java
@@ -1,0 +1,31 @@
+package com.whatpl.global.security.handler;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.whatpl.global.exception.ErrorCode;
+import com.whatpl.global.exception.ErrorResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+@RequiredArgsConstructor
+public class NoAuthorizationHandler implements AccessDeniedHandler {
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+        ErrorResponse errorResponse = ErrorResponse.of(ErrorCode.NO_AUTHORIZATION);
+        response.setStatus(HttpStatus.FORBIDDEN.value());
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
+        objectMapper.writeValue(response.getWriter(), errorResponse);
+    }
+}
