--- src/main/java/com/whatpl/jwt/JwtProperties.java
@@ -1,46 +1,0 @@
-package com.whatpl.jwt;
-
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.validation.annotation.Validated;
-
-import javax.crypto.SecretKey;
-
-@Validated
-@Getter
-@AllArgsConstructor
-@ConfigurationProperties(prefix = "jwt")
-public class JwtProperties {
-
-    @NotBlank
-    private String secret;
-
-    @NotBlank
-    private String header;
-
-    @NotBlank
-    private String tokenType;
-
-    @Min(60_000L)
-    @Max(86_400_000L)
-    private long accessExpirationTime;
-
-    @Min(86_400_000L)
-    @Max(2_592_000_000L)
-    private long refreshExpirationTime;
-
-    private SecretKey secretKey;
-
-    @PostConstruct
-    private void init() {
-        this.secretKey = Keys.hmacShaKeyFor(Decoders.BASE64URL.decode(secret));
-    }
-}
-
