--- src/main/java/com/whatpl/jwt/JwtResponse.java
@@ -1,20 +1,0 @@
-package com.whatpl.jwt;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class JwtResponse {
-
-    private String accessToken;
-    private String refreshToken;
-
-    @Builder
-    public JwtResponse(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-}
