--- src/main/java/com/whatpl/security/handler/NoAuthenticationHandler.java
@@ -1,31 +1,0 @@
-package com.whatpl.security.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.whatpl.exception.ErrorCode;
-import com.whatpl.exception.ErrorResponse;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-
-@RequiredArgsConstructor
-public class NoAuthenticationHandler implements AuthenticationEntryPoint {
-
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        ErrorResponse errorResponse = ErrorResponse.of(ErrorCode.NO_AUTHENTICATION);
-        response.setStatus(HttpStatus.UNAUTHORIZED.value());
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setCharacterEncoding(StandardCharsets.UTF_8.displayName());
-        objectMapper.writeValue(response.getWriter(), errorResponse);
-    }
-}
