--- src/main/java/com/whatpl/security/repository/CookieOAuth2AuthorizationRequestRepository.java
@@ -1,64 +1,0 @@
-package com.whatpl.security.repository;
-
-import com.whatpl.util.CookieUtils;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.oauth2.client.web.AuthorizationRequestRepository;
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames;
-import org.springframework.util.Assert;
-
-public class CookieOAuth2AuthorizationRequestRepository
-        implements AuthorizationRequestRepository<OAuth2AuthorizationRequest> {
-
-    private static final String OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME = "OAUTH2_AUTH_REQUEST";
-    private static final int OAUTH2_COOKIE_EXPIRATION_SECONDS = 30 * 60;
-
-    @Override
-    public OAuth2AuthorizationRequest loadAuthorizationRequest(HttpServletRequest request) {
-        Assert.notNull(request, "request cannot be null");
-        String stateParameter = getStateParameter(request);
-        if (stateParameter == null) {
-            return null;
-        }
-        OAuth2AuthorizationRequest authorizationRequest = getAuthorizationRequest(request);
-        return (authorizationRequest != null && stateParameter.equals(authorizationRequest.getState()))
-                ? authorizationRequest : null;
-    }
-
-    @Override
-    public void saveAuthorizationRequest(OAuth2AuthorizationRequest authorizationRequest, HttpServletRequest request, HttpServletResponse response) {
-        Assert.notNull(request, "request cannot be null");
-        Assert.notNull(response, "response cannot be null");
-        if (authorizationRequest == null) {
-            removeAuthorizationRequest(request, response);
-            return;
-        }
-        String state = authorizationRequest.getState();
-        Assert.hasText(state, "authorizationRequest.state cannot be empty");
-        CookieUtils.addCookie(response,
-                OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME,
-                CookieUtils.serialize(authorizationRequest),
-                OAUTH2_COOKIE_EXPIRATION_SECONDS);
-    }
-
-    @Override
-    public OAuth2AuthorizationRequest removeAuthorizationRequest(HttpServletRequest request, HttpServletResponse response) {
-        Assert.notNull(response, "response cannot be null");
-        OAuth2AuthorizationRequest authorizationRequest = loadAuthorizationRequest(request);
-        if (authorizationRequest != null) {
-            CookieUtils.deleteCookie(request, response, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME);
-        }
-        return authorizationRequest;
-    }
-
-    private String getStateParameter(HttpServletRequest request) {
-        return request.getParameter(OAuth2ParameterNames.STATE);
-    }
-
-    private OAuth2AuthorizationRequest getAuthorizationRequest(HttpServletRequest request) {
-        return CookieUtils.getCookie(request, OAUTH2_AUTHORIZATION_REQUEST_COOKIE_NAME)
-                .map(cookie -> CookieUtils.deserialize(cookie, OAuth2AuthorizationRequest.class))
-                .orElse(null);
-    }
-}
