+++ src/test/java/com/whatpl/oauth2/OAuth2ApiTest.java
@@ -1,10 +1,10 @@
 package com.whatpl.oauth2;
 
 import com.whatpl.account.AccountService;
-import com.whatpl.security.config.SecurityConfig;
-import com.whatpl.jwt.JwtProperties;
-import com.whatpl.jwt.JwtService;
-import com.whatpl.util.CookieUtils;
+import com.whatpl.global.config.SecurityConfig;
+import com.whatpl.global.jwt.JwtProperties;
+import com.whatpl.global.jwt.JwtService;
+import com.whatpl.global.util.CookieUtils;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
