+++ src/test/java/com/whatpl/security/jwt/JwtServiceTest.java
@@ -2,14 +2,14 @@ package com.whatpl.security.jwt;
 
 import com.whatpl.account.Account;
 import com.whatpl.account.AccountRepository;
-import com.whatpl.exception.BizException;
-import com.whatpl.exception.ErrorCode;
-import com.whatpl.jwt.JwtProperties;
-import com.whatpl.jwt.JwtResponse;
-import com.whatpl.jwt.JwtService;
-import com.whatpl.redis.RedisService;
-import com.whatpl.security.domain.AccountPrincipal;
-import com.whatpl.security.domain.OAuth2UserInfo;
+import com.whatpl.global.exception.BizException;
+import com.whatpl.global.exception.ErrorCode;
+import com.whatpl.global.jwt.JwtProperties;
+import com.whatpl.global.jwt.JwtResponse;
+import com.whatpl.global.jwt.JwtService;
+import com.whatpl.global.redis.RedisService;
+import com.whatpl.global.security.domain.AccountPrincipal;
+import com.whatpl.global.security.domain.OAuth2UserInfo;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.io.Decoders;
